requirements.txt
Django>=2.1.3,<3.0.7
djangorestframework>=3.9.0,<3.11.0
